requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
supabase
python-decouple
psycopg2-binary
Pillow
python-dotenv